# Add trivial type conversion for Euler types