This PR includes the changes in #143 to pass the CI, so please merge the PR first. :pray:

# Before this PR
```julia
julia> RotXY(RotX(10))
ERROR: Cannot construct a two-axis rotation from a matrix
Stacktrace:
 [1] error(s::String)
   @ Base ./error.jl:33
 [2] RotXY
   @ ~/.julia/dev/Rotations/src/euler_types.jl:241 [inlined]
 [3] RotXY(a::RotX{Float64})
   @ StaticArrays ~/.julia/packages/StaticArrays/rdb0l/src/convert.jl:5
 [4] top-level scope
   @ REPL[3]:1

julia> RotXYZ(RotX(10))
3×3 RotXYZ{Float64} with indices SOneTo(3)×SOneTo(3)(-2.56637, 0.0, -0.0):
 1.0   0.0        0.0
 0.0  -0.839072   0.544021
 0.0  -0.544021  -0.839072

julia> @benchmark RotXYZ(RotX(10))
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     39.974 ns (0.00% GC)
  median time:      40.712 ns (0.00% GC)
  mean time:        41.134 ns (0.00% GC)
  maximum time:     60.912 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     991
```

# After this PR
```julia
julia> RotXY(RotX(10))
3×3 RotXY{Float64} with indices SOneTo(3)×SOneTo(3)(10.0, 0.0):
  1.0   0.0        0.0
 -0.0  -0.839072   0.544021
  0.0  -0.544021  -0.839072

julia> RotXYZ(RotX(10))
3×3 RotXYZ{Float64} with indices SOneTo(3)×SOneTo(3)(10.0, 0.0, 0.0):
  1.0  -0.0        0.0
 -0.0  -0.839072   0.544021
  0.0  -0.544021  -0.839072

julia> @benchmark RotXYZ(RotX(10))
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     0.020 ns (0.00% GC)
  median time:      0.020 ns (0.00% GC)
  mean time:        0.023 ns (0.00% GC)
  maximum time:     8.646 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000
```

# Note
In some cases, the type conversion can be non-trivial. For example, `RotX(t)` can be conveted to two ways `RotXYX(t,0,0)` and `RotXYX(0,0,t)`. With this PR, the first `RotXYX(t,0,0)` will be choosen.

```julia
julia> RotXYX(RotX(10))
3×3 RotXYX{Float64} with indices SOneTo(3)×SOneTo(3)(10.0, 0.0, 0.0):
  1.0   0.0        0.0
 -0.0  -0.839072   0.544021
  0.0  -0.544021  -0.839072

julia> RotZYZ(RotZ(42))
3×3 RotZYZ{Float64} with indices SOneTo(3)×SOneTo(3)(42.0, 0.0, 0.0):
 -0.399985   0.916522  -0.0
 -0.916522  -0.399985  -0.0
 -0.0        0.0        1.0
```