@hyrodium 2021-04-09
I forgot to add the following conversion, so I'll change this PR to draft.
```julia
julia> RotXYX(RotZ(103))
3×3 RotXYX{Float64} with indices SOneTo(3)×SOneTo(3)(1.5708, 2.46904, -1.5708):
 -0.782231     -0.622989     -1.00184e-16
  0.622989     -0.782231     -6.456e-17
 -3.81471e-17  -1.12914e-16   1.0

julia> RotXYX(π/2,103,-π/2)
3×3 RotXYX{Float64} with indices SOneTo(3)×SOneTo(3)(1.5708, 103.0, -1.5708):
 -0.782231     -0.622989    3.81471e-17
  0.622989     -0.782231    1.0913e-16
 -3.81471e-17   1.0913e-16  1.0
```