@hyrodium 2021-04-10
I've added method for `RotXYX(RotZ(103))`, `RotXZX(RotY(103))`, `RotYXY(RotZ(103))`, and etc.
Here's the benchmark:
# Before This PR
```julia
julia> RotXYX(RotZ(103))
3×3 RotXYX{Float64} with indices SOneTo(3)×SOneTo(3)(1.5708, 2.46904, -1.5708):
 -0.782231     -0.622989     -1.00184e-16
  0.622989     -0.782231     -6.456e-17
 -3.81471e-17  -1.12914e-16   1.0

julia> @benchmark RotXYX(RotZ(103))
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     42.892 ns (0.00% GC)
  median time:      44.726 ns (0.00% GC)
  mean time:        44.893 ns (0.00% GC)
  maximum time:     75.555 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     990
```

# Afer this PR
```julia
julia> RotXYX(RotZ(103))
3×3 RotXYX{Float64} with indices SOneTo(3)×SOneTo(3)(1.5708, 103.0, -1.5708):
 -0.782231     -0.622989    3.81471e-17
  0.622989     -0.782231    1.0913e-16
 -3.81471e-17   1.0913e-16  1.0

julia> @benchmark RotXYX(RotZ(103))
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     0.838 ns (0.00% GC)
  median time:      0.978 ns (0.00% GC)
  mean time:        1.109 ns (0.00% GC)
  maximum time:     10.266 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000
```

And I also added tests for 1-axis -> 2-axis.
I think this PR is ready, so I'll remove the draft state:tada: